Remove `lingo init` command
Requirement: `initRepo()` runs silently at the start of each `lingo review` run.

  * Move necessary logic from `app/commands/init.go` to `app/commands/review.go`

  * Change return type of `initRepo()` from `(string, string, error)` → `error`, as the repoName and repoOwner no longer need to be evaluated by the caller.

  * Delete `app/commands/init.go`
