+++ app/commands/review.go
@@ -10,7 +10,15 @@ import (
 
 	"github.com/codelingo/lingo/app/util"
 
+	"os"
+	"path/filepath"
+	"strconv"
+	"strings"
+
 	"github.com/codegangsta/cli"
+
+	"github.com/codelingo/lingo/vcs"
+	"github.com/codelingo/lingo/vcs/backing"
 )
 
 func init() {
@@ -52,6 +60,14 @@ func reviewAction(ctx *cli.Context) {
 }
 
 func reviewCMD(ctx *cli.Context) (string, error) {
+	if err := initRepo(ctx); err != nil {
+		// TODO(waigani) use error types
+		// Note: Prior repo init is a valid state.
+		if !strings.Contains(err.Error(), "already exists") {
+			return "", errors.Trace(err)
+		}
+	}
+
 	dotlingo, err := readDotLingo(ctx)
 	if err != nil {
 		return "", errors.Trace(err)
@@ -84,4 +100,69 @@ func readDotLingo(ctx *cli.Context) (string, error) {
 		}
 	}
 	return string(dotlingo), nil
+}
+
+// TODO(waigani) start ingesting repo as soon as it's inited
+
+func initRepo(ctx *cli.Context) error {
+
+	repo := vcs.New(backing.Git)
+	authCfg, err := util.AuthConfig()
+	if err != nil {
+		return errors.Trace(err)
+	}
+
+	if err = repo.AssertNotTracked(); err != nil {
+		// TODO (benjamin-rood): Check the error type
+		return errors.Trace(err)
+	}
+
+	// TODO(waigani) Try to get owner and name from origin remote first.
+
+	// get the repo owner name
+	repoOwner, err := authCfg.Get("gitserver.user.username")
+	if err != nil {
+		return errors.Trace(err)
+	}
+
+	// get the repo name, default to working directory name
+	dir, err := os.Getwd()
+	if err != nil {
+		return errors.Trace(err)
+	}
+
+	repoName := filepath.Base(dir)
+
+	// TODO(benjamin-rood) Check if repo name and contents are identical.
+	// If, so this should be a no-op and only the remote needs to be set.
+	// ensure creation of distinct remote.
+	repoName, err = createRepo(repo, repoName)
+	if err != nil {
+		return errors.Trace(err)
+	}
+	_, _, err = repo.SetRemote(repoOwner, repoName)
+	return err
+}
+
+func createRepo(repo backing.Repo, name string) (string, error) {
+	if err := repo.CreateRemote(name); err != nil {
+		if strings.Contains(err.Error(), "already exists") {
+			parts := strings.Split(name, "-")
+			num := parts[len(parts)-1]
+
+			// We ignore the error here because the only case in which Atoi
+			// would error is if the name had not yet been appended with -n.
+			// In this case, n will be set to zero which is what we require.
+			n, _ := strconv.Atoi(num)
+			if n != 0 {
+				// Need to remove existing trailing number where present,
+				// otherwise the repoName only appends rather than replaces
+				// and will produce names of the pattern "myPkg-1-2-...-n-n+1".
+				name = strings.TrimSuffix(name, "-"+num)
+			}
+			return createRepo(repo, fmt.Sprintf("%s-%d", name, n+1))
+		}
+		return "", err
+	}
+	return name, nil
 }
